# docs/swagger.yaml
basePath: /
definitions:
...
  input.AppConfigSaveConfigReq:
    properties:
      login_bg:
        description: 登录页背景图路径（留空则根据主题色生成背景）
        example: ""
        type: string
      login_slogan:
        description: 登录页副标题 / Slogan
        example: ""
        type: string
      login_title:
        description: 登录页主标题
        example: 积分商城
        type: string
      logo:
        description: 网站 Logo 路径（登录页等场景展示）
        example: ""
        type: string
      register:
        description: 是否允许用户自助注册, 0-禁止, 1-允许
        enum:
        - "0"
        - "1"
        example: "1"
        type: string
      site_background_color:
        description: PWA 启动页背景色
        example: '#f5f6f8'
        type: string
      site_description:
        description: 站点说明（PWA 应用描述）
        example: 这是xxxxx的积分商城
        type: string
      site_icon:
        description: 网站图标路径（标签页 / 桌面应用图标）
        example: ""
        type: string
      site_name:
        description: 站点名称（浏览器标签栏 / PWA 安装后名称）
        example: 积分商城
        type: string
      site_theme_color:
        description: 网站主题色
        example: '#965bff'
        type: string
    required:
    - register
    type: object
  input.AppConfigSaveOssConfigReq:
    properties:
      oss_access_key_id:
        description: 阿里云 AccessKey ID
        example: xxxxxxxxxxxxx
        type: string
      oss_access_key_secret:
        description: 阿里云 AccessKey Secret
        example: xxxxxxxxxxxxx
        type: string
      oss_bucket:
        description: 目标 bucket 名
        example: xxxxx
        type: string
      oss_endpoint:
        description: 完整 OSS 地址
        example: https://oss-cn-hangzhou.aliyuncs.com
        type: string
    type: object
  input.FeedbackDetailsReq:
    properties:
      id:
        description: 投诉ID
        example: 1
        minimum: 1
        type: integer
    required:
    - id
    type: object
  input.FeedbackListPageReq:
    properties:
      pageNo:
        description: 页码
        example: 1
        type: integer
      pageSize:
        description: 每页展示条数
        example: 20
        type: integer
      sortField:
        description: 排序字段，取投诉表列名或 uid/uname
        example: created_at
        type: string
      sortOrder:
        description: 排序方向 ascend/descend
        enum:
        - ascend
        - descend
        example: descend
        type: string
      uid:
        description: 用户UID，精确匹配
        example: 54272611
        type: integer
      uname:
        description: 用户昵称，模糊搜索
        example: 哎呀又胖啦
        type: string
    required:
    - pageNo
    - pageSize
    type: object
  input.FeedbackSubmitReq:
    properties:
      contact:
        description: 联系方式（手机号 / 邮箱 / QQ 等，便于平台回访）
        example: "18888888888"
        maxLength: 100
        type: string
      content:
        description: 投诉内容
        example: 主播发布违规内容，请核查
        maxLength: 5000
        type: string
      type:
        description: 问题类型
        example: 直播问题
        maxLength: 100
        type: string
    required:
    - contact
    - content
    - type
    type: object
...
  resp.AppConfigDataResp:
    properties:
      login_bg:
        description: 登录页背景图路径（留空则根据主题色生成背景）
        example: ""
        type: string
      login_slogan:
        description: 登录页副标题 / Slogan
        example: ""
        type: string
      login_title:
        description: 登录页主标题
        example: 积分商城
        type: string
      logo:
        description: 网站 Logo 路径（登录页等场景展示）
        example: https://cdn.hejunjie.life/avatars/shop.png
        type: string
      oss_access_key_id:
        description: 阿里云 AccessKey ID
        example: xxxxxxxxxxxxx
        type: string
      oss_access_key_secret:
        description: 阿里云 AccessKey Secret
        example: xxxxxxxxxxxxx
        type: string
      oss_bucket:
        description: 目标 bucket 名
        example: xxxxx
        type: string
      oss_endpoint:
        description: 完整 OSS 地址
        example: https://oss-cn-hangzhou.aliyuncs.com
        type: string
      register:
        description: 是否允许用户自助注册, 0-禁止, 1-允许
        example: "1"
        type: string
      site_background_color:
        description: PWA 启动页背景色
        example: '#f5f6f8'
        type: string
      site_description:
        description: 站点说明（PWA 应用描述）
        example: 这是xxxxx的积分商城
        type: string
      site_icon:
        description: 网站图标路径（标签页 / 桌面应用图标）
        example: https://cdn.hejunjie.life/avatars/shop.png
        type: string
      site_name:
        description: 站点名称（浏览器标签栏 / PWA 安装后名称）
        example: 积分商城
        type: string
      site_theme_color:
        description: 网站主题色
        example: '#965bff'
        type: string
    type: object
  resp.AppPublicKeyResp:
    properties:
      key_id:
        description: 公钥标识（公钥内容 sha256 前 16 位 hex，用于前端验签与密钥轮换识别）
        example: 3f2ab8d0e1c4a9b7
        type: string
      public_key:
        description: RSA 公钥（SPKI DER 的 base64 编码，前端 atob 后 importKey("spki") 使用）
        example: MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8A...
        type: string
      sign:
        description: HMAC-SHA256 签名（对 "pubkey:"+key_id+public_key+timestamp 计算，hex
          编码）
        example: a1b2c3d4e5f60718293a4b5c6d7e8f90...
        type: string
      timestamp:
        description: 签名生成时间戳（Unix 秒，前端校验时间窗口）
        example: 1724716800
        type: integer
    type: object
  resp.AppShopLoginConfigResp:
    properties:
      login_bg:
        example: ""
        type: string
      logo:
        example: https://cdn.hejunjie.life/avatars/shop.png
        type: string
      register:
        example: false
        type: boolean
      slogan:
        example: 登录后可兑换积分好礼
        type: string
      title:
        example: 积分商城
        type: string
    type: object
  resp.AppShopManifestIcon:
    properties:
      purpose:
        description: 用途（any / maskable）
        example: any
        type: string
      sizes:
        description: 尺寸
        example: 192x192
        type: string
      src:
        description: 图标地址
        example: https://cdn.example.com/icon-192.png
        type: string
      type:
        description: 图片类型
        example: image/png
        type: string
    type: object
  resp.AppShopManifestResp:
    properties:
      apple_touch_icon:
        description: iOS 添加到主屏的图标（180x180 PNG）
        example: https://cdn.example.com/icon-180.png
        type: string
      background_color:
        description: 启动屏背景色
        example: '#ffffff'
        type: string
      description:
        description: 应用描述
        example: 积分商城的描述
        type: string
      display:
        description: 显示模式
        example: standalone
        type: string
      favicon:
        description: 浏览器标签页小图标
        example: https://cdn.example.com/favicon.svg
        type: string
      icons:
        description: 安装图标列表
        items:
          $ref: '#/definitions/resp.AppShopManifestIcon'
        type: array
      name:
        description: 应用名称（全称，用于页面标题与安装名称）
        example: 积分商城
        type: string
      scope:
        description: 作用域
        example: /shop/
        type: string
      short_name:
        description: 应用短名称（空间不足时展示）
        example: 商城
        type: string
      start_url:
        description: 启动地址
        example: /shop/
        type: string
      theme_color:
        description: 主题色（浏览器地址栏 / 窗口标题栏颜色）
        example: '#ffffff'
        type: string
    type: object
  resp.AppShopThemeColorResp:
    properties:
      color:
        example: '#ffffff'
        type: string
    type: object
  resp.FeedbackDetailsResp:
    properties:
      contact:
        description: 联系方式
        example: "18888888888"
        type: string
      content:
        description: 投诉内容
        example: 主播发布违规内容，请核查
        type: string
      created_at:
        description: 投诉时间
        example: xxxx-xx-xx xx:xx:xx
        type: string
      id:
        description: 投诉ID
        example: 1
        type: integer
      type:
        description: 问题类型
        example: 直播问题
        type: string
      uid:
        description: 用户UID
        example: 54272611
        type: integer
      uname:
        description: 用户昵称
        example: 哎呀又胖啦
        type: string
      user_id:
        description: 用户内部ID
        example: 1
        type: integer
    type: object
  resp.FeedbackListPageItem:
    properties:
      contact:
        description: 联系方式
        example: "18888888888"
        type: string
      created_at:
        description: 投诉时间
        example: xxxx-xx-xx xx:xx:xx
        type: string
      id:
        description: 投诉ID
        example: 1
        type: integer
      type:
        description: 问题类型
        example: 直播问题
        type: string
      uid:
        description: 用户UID
        example: 54272611
        type: integer
      uname:
        description: 用户昵称
        example: 哎呀又胖啦
        type: string
      user_id:
        description: 用户内部ID
        example: 1
        type: integer
    type: object
  resp.FeedbackListPageResp:
    properties:
      pageData:
        description: 当前页码数据
        items:
          $ref: '#/definitions/resp.FeedbackListPageItem'
        type: array
      total:
        description: 总计条数
        example: 100
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应code
        example: 0
        type: integer
      data:
        description: 响应数据
      msg:
        description: 响应信息
        example: success
        type: string
    type: object
info:
  contact:
    email: junjie.he.925@gmail.com
    name: 何俊杰
  description: BiliLive Assistant 系统接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BiliLive Assistant API
  version: "1.0"
paths:
...
  /api/admin/appconfig/data:
    post:
      description: 获取后台可编辑的全部 App 配置（站点信息、颜色、图标、注册开关、登录页配置），用于回填管理页表单
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AppConfigDataResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取 App 配置
      tags:
      - App配置
  /api/admin/appconfig/oss_save:
    post:
      description: 整体覆盖保存 OSS 相关配置并刷新缓存立即生效，可为空表示清除对应配置
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: App 配置参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AppConfigSaveOssConfigReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 保存 OSS 配置
      tags:
      - App配置
  /api/admin/appconfig/save:
    post:
      description: 整体覆盖保存基本 App 配置并刷新缓存立即生效，可为空表示清除对应配置
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: App 配置参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AppConfigSaveConfigReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 保存 App 基础配置
      tags:
      - App配置
...
  /api/admin/feedback/details:
    post:
      description: 根据投诉ID返回投诉全部字段（含投诉正文），以及投诉用户的 uid/uname
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.FeedbackDetailsReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.FeedbackDetailsResp'
              type: object
      security:
      - BearerAuth: []
      summary: 后台获取投诉详情
      tags:
      - 投诉管理
  /api/admin/feedback/list:
    post:
      description: 分页查询用户提交的投诉/反馈，联查用户表返回用户 UID 与昵称；支持按 UID 精确、按昵称模糊筛选。列表不下发投诉正文，正文请调详情接口
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.FeedbackListPageReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.FeedbackListPageResp'
              type: object
      security:
      - BearerAuth: []
      summary: 后台分页查询投诉列表
      tags:
      - 投诉管理
...
  /api/shop/feedback/submit:
    post:
      description: 当前登录用户提交一条投诉/反馈（问题类型、内容、联系方式），提交后由平台侧处理；联系方式仅用于平台回访
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.FeedbackSubmitReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 提交用户投诉/反馈
      tags:
      - 移动端
...
  /api/shop/login:
    get:
      description: 获取登录页配置（注册开关、Logo、背景图、标题、Slogan）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AppShopLoginConfigResp'
              type: object
      summary: 获取登录页面配置信息
      tags:
      - 移动端
  /api/shop/manifest:
    get:
      description: 获取 App 的 Manifest 信息，用于前端构建 PWA 应用
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AppShopManifestResp'
              type: object
      summary: 获取 App 的 Manifest 信息
      tags:
      - 移动端
...
  /api/shop/public-key:
    get:
      description: 获取用于前端 RSA-OAEP 加密的 RSA 公钥（SPKI DER 的 base64），并附带 HMAC 签名供前端验签
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AppPublicKeyResp'
              type: object
      summary: 获取 RSA 公钥（带 HMAC 验签）
      tags:
      - 移动端
  /api/shop/theme-color:
    get:
      description: 获取 App 主题色，用于前端构建样式
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AppShopThemeColorResp'
              type: object
      summary: 获取 App 主题色
      tags:
      - 移动端
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"